sql/113: correct

+++ sql/113-create-table3.sql
@@ -6,8 +6,8 @@
 CREATE TABLE weather (
     id INTEGER PRIMARY KEY NOT NULL,
     time_of_reading VARCHAR(50) NOT NULL,
-    high INTEGER NOT NULL,
-    low INTEGER NOT NULL,
+    high FLOAT NOT NULL,
+    low FLOAT NOT NULL,
     "signature" VARCHAR(50),
     comment VARCHAR(50),
 );
